+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [2.1.0] - TODO
 ### Fixed
-- Replace unsafe innerHTML assignment in ButtonScroll with safe DOM manipulation.
+- Replace unsafe innerHTML and insertAdjacentHTML usage with safe DOM manipulation when displaying information in the add-on’s popup.
 
 ## [2.0.0] - 20260620
 ### Added